requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
boto3==1.34.76
botocore==1.34.76
Django==4.2.11
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
mysqlclient==2.2.4
pillow==10.2.0
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
